+++ src/Codeception/TestCase/Cest.php
@@ -61,6 +61,9 @@ class Cest extends \Codeception\TestCase implements
             $this->executeTestMethod($I);
             $this->executeAfter($this->testMethod, $I);
         } catch (\Exception $e) {
+            if (is_callable([$this->testClassInstance, '_failed'])) {
+                $this->testClassInstance->_failed($I);
+            }
             // fails and errors are now handled by Codeception\PHPUnit\Listener
             throw $e;
         }
@@ -70,7 +73,7 @@ class Cest extends \Codeception\TestCase implements
 
     protected function executeBefore($testMethod, $I)
     {
-        if (method_exists($this->testClassInstance, '_before')) {
+        if (is_callable([$this->testClassInstance, '_before'])) {
             $this->testClassInstance->_before($I);
         }
 
@@ -91,7 +94,7 @@ class Cest extends \Codeception\TestCase implements
             }
         }
 
-        if (method_exists($this->testClassInstance, '_after')) {
+        if (is_callable([$this->testClassInstance, '_after'])) {
             $this->testClassInstance->_after($I);
         }
     }