+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### 2.0.10
 
+* Added '_failed' hook for Cest tests. Fixes #1660 *2015-02-02*
 * [REST] fixed setting Host header. Issue #1650 *2015-02-02*
 * [Laravel4] Disconnecting from database after each test to prevent Too many connections #1665 by @mnabialek *2015-02-02*
 * [Symfony2] Fixed kernel reusage in #1656 by @hacfi *2015-02-01*